Fix uniform random distribution bugs and rename files
generate_random.py renamed to random_generator.py and test_generate_random.py renamed to test_random_generator.py. Also, improved uniform random distribution tests and implementation.
